integration_engine/requirements.txt
# Core dependencies
aiofiles>=0.8.0
asyncio>=3.4.3
pydantic==1.10.0
PyYAML>=6.0
redis>=4.5.0
python-dotenv>=0.21.0
fastapi>=0.85.0
uvicorn>=0.19.0
python-multipart>=0.0.5
httpx>=0.23.0
pytest>=7.2.0
pytest-asyncio>=0.20.0
python-dateutil>=2.8.2
pytz>=2022.6
Jinja2>=3.1.2
jsonschema>=4.16.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
prometheus-client>=0.15.0
